# src/main/resources/swagger/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Patient Data API
  license:
    name: MIT
servers:
  - url: http://localhost:8080/
paths:
  /login:
    post:
      summary: "User login"
      tags:
        - auth
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: "Successful login"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '403':
          description: "Invalid login credentials"

  /patients:
    get:
      summary: List all patients
      operationId: listPatients
      tags:
        - patients
      parameters:
        - name: page
          in: query
          description: The page number to retrieve (default 0)
          required: false
          schema:
            type: integer
            default: 0
            format: int32
        - name: limit
          in: query
          description: How many items to return at one time (default 10, max 100)
          required: false
          schema:
            type: integer
            default: 10
            format: int32
      responses:
        '200':
          description: A paged array of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientResponse"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Register a new patient
      operationId: createPatient
      tags:
        - patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRequest"
      responses:
        '201':
          description: Patient successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResponse"
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /patients/{patientId}:
    get:
      summary: Get information for a specific patient
      operationId: showPatientById
      tags:
        - patients
      parameters:
        - name: patientId
          in: path
          required: true
          description: The ID of the patient to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response for a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResponse"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update information for a specific patient
      operationId: updatePatient
      tags:
        - patients
      parameters:
        - name: patientId
          in: path
          required: true
          description: The ID of the patient to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRequest"
      responses:
        '200':
          description: Patient successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResponse"
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a patient record
      operationId: deletePatient
      tags:
        - patients
      parameters:
        - name: patientId
          in: path
          required: true
          description: The ID of the patient to delete
          schema:
            type: string
      responses:
        '204':
          description: Patient successfully deleted
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
    PatientRequest:
      type: object
      required:
        - firstName
        - lastName
        - gender
        - age
        - email
        - phoneNumber
      properties:
        firstName:
          type: string
          description: First name of the patient
          example: John
        lastName:
          type: string
          description: Last name of the patient
          example: Doe
        gender:
          type: string
          description: Gender of the patient
          example: male
        age:
          type: integer
          minimum: 0
          maximum: 120
          description: Age of the patient
          example: 30
        email:
          type: string
          format: email
          description: Patient’s email address
          example: john.doe@example.com
        phoneNumber:
          type: string
          pattern: "^\\+?[0-9. ()-]{7,}$"
          description: Contact phone number of the patient
          example: "+1234567890"
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          description: Patient id
        firstName:
          type: string
          description: First name of the patient
          example: John
        lastName:
          type: string
          description: Last name of the patient
          example: Doe
        gender:
          type: string
          description: Gender of the patient
          example: male
        age:
          type: integer
          description: Age of the patient
          example: 30
        email:
          type: string
          description: Patient’s email address
          example: john.doe@example.com
        phoneNumber:
          type: string
          description: Contact phone number of the patient
          example: "+1234567890"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Detailed error message
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: string
            description: Validation error messages keyed by field name
    UserNotFoundError:
      type: object
      properties:
        error:
          type: string
          description: Error message for user not found